chore(stepper): Move stepper to ui/react-components
