+++ packages/components/src/Stepper/Stepper.component.scss
