+++ packages/components/src/Stepper/Stepper.stories.js
@@ -1,17 +1,9 @@
 import React, { useState } from 'react';
-import Action from '@talend/react-components/lib/Actions/Action';
-import Stepper from '@talend/react-components/lib/Stepper';
-
-// eslint-disable-next-line
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-
-import { StepperConstants } from '../src';
-
-import './style.scss';
-
-const stories = storiesOf('Stepper', module);
-
+import Action from '../Actions/Action';
+import Stepper, { LOADING_STEP_STATUSES } from './Stepper.component';
+const stories = storiesOf('Messaging & Communication/Stepper', module);
 const title = 'Sample processing...';
 
 function renderActions(isInError) {
@@ -30,15 +22,15 @@ function renderActions(isInError) {
 stories
 	.add('Stepper default', () => {
 		const steps = [
-			{ label: 'Fetch Sample', status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS },
+			{ label: 'Fetch Sample', status: LOADING_STEP_STATUSES.SUCCESS },
 			{
 				label: 'Global Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.LOADING,
+				status: LOADING_STEP_STATUSES.LOADING,
 			},
-			{ label: 'Flattening', status: StepperConstants.LOADING_STEP_STATUSES.LOADING },
+			{ label: 'Flattening', status: LOADING_STEP_STATUSES.LOADING },
 			{
 				label: 'Column Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.PENDING,
+				status: LOADING_STEP_STATUSES.PENDING,
 			},
 		];
 		return <Stepper steps={steps} title={title} />;
@@ -47,18 +39,18 @@ stories
 		const steps = [
 			{
 				label: 'Fetch Sample',
-				status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+				status: LOADING_STEP_STATUSES.SUCCESS,
 				message: { label: 'Everything is fine 🔥🐶' },
 			},
 			{
 				label: 'Global Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.FAILURE,
+				status: LOADING_STEP_STATUSES.FAILURE,
 				message: { label: "We couldn't connect to the remote engine" },
 			},
-			{ label: 'Flattening', status: StepperConstants.LOADING_STEP_STATUSES.ABORTED },
+			{ label: 'Flattening', status: LOADING_STEP_STATUSES.ABORTED },
 			{
 				label: 'Column Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.ABORTED,
+				status: LOADING_STEP_STATUSES.ABORTED,
 			},
 		];
 		return <Stepper steps={steps} title={title} renderActions={renderActions} />;
@@ -70,15 +62,15 @@ stories
 	))
 	.add('Stepper successful', () => {
 		const defaultSteps = [
-			{ label: 'Fetch Sample', status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS },
+			{ label: 'Fetch Sample', status: LOADING_STEP_STATUSES.SUCCESS },
 			{
 				label: 'Global Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+				status: LOADING_STEP_STATUSES.SUCCESS,
 			},
-			{ label: 'Flattening', status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS },
+			{ label: 'Flattening', status: LOADING_STEP_STATUSES.SUCCESS },
 			{
 				label: 'Column Quality',
-				status: StepperConstants.LOADING_STEP_STATUSES.LOADING,
+				status: LOADING_STEP_STATUSES.LOADING,
 			},
 		];
 
@@ -89,19 +81,19 @@ stories
 				setSteps([
 					{
 						label: 'Fetch Sample',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Global Quality',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Flattening',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Column Quality',
-						status: StepperConstants.LOADING_STEP_STATUSES.LOADING,
+						status: LOADING_STEP_STATUSES.LOADING,
 					},
 				]);
 			};
@@ -110,19 +102,19 @@ stories
 				setSteps([
 					{
 						label: 'Fetch Sample',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Global Quality',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Flattening',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 					{
 						label: 'Column Quality',
-						status: StepperConstants.LOADING_STEP_STATUSES.SUCCESS,
+						status: LOADING_STEP_STATUSES.SUCCESS,
 					},
 				]);
 			};
