+++ packages/components/src/Stepper/StepperTransition.component.js
