+++ packages/components/src/Stepper/index.js
@@ -1,0 +1,9 @@
+import Stepper, {
+	LOADING_STEP_STATUSES,
+	isStepsLoading,
+	isAllSuccessful,
+	isErrorInSteps,
+} from './Stepper.component';
+
+export { LOADING_STEP_STATUSES, isErrorInSteps, isAllSuccessful, isStepsLoading };
+export default Stepper;
