--- packages/stepper/src/components/constant.js
@@ -1,3 +1,0 @@
-const I18N_NAMESPACE = 'tui-stepper';
-
-export default I18N_NAMESPACE;
