+++ packages/stepper/src/index.js
@@ -1,7 +1,7 @@
+import { isAllSuccessful, isStepsLoading, isErrorInSteps } from '@talend/react-components/lib/Stepper';
 import stepperReducer from './service/Stepper.reducer';
 import { initStepper, proceedLoadingEvent, removeStepper } from './service/Stepper.actions';
 import { getStepsForResource, isResourceLoading } from './service/Stepper.selectors';
-import { isAllSuccessful, isStepsLoading, isErrorInSteps } from './service/Stepper.utils';
 import Stepper from './containers/Stepper.connect';
 import { STATE_KEY, LOADING_STEP_STATUSES } from './Stepper.constants';
