+++ packages/stepper/src/service/Stepper.selectors.js
@@ -1,6 +1,7 @@
 import get from 'lodash/get';
+import { isStepsLoading } from '@talend/react-components/lib/Stepper';
 import { STATE_KEY } from '../Stepper.constants';
-import { getStepperKey, isStepsLoading } from './Stepper.utils';
+import { getStepperKey } from './Stepper.utils';
 
 export function getStepsForResource(store, resourceType, resourceId) {
 	return get(store, [STATE_KEY, getStepperKey({ resourceType, resourceId }), 'steps'], []);
